+++ src/test/unit/common/proxy.test.ts
@@ -11,7 +11,8 @@ suite('Proxy', () => {
   const host = 'my.proxy.com';
   const port = 8080;
   const auth = 'user:password';
-  const proxy = `https://${auth}@${host}:${port}`;
+  const protocol = 'https:';
+  const proxy = `${protocol}//${auth}@${host}:${port}`;
   const proxyStrictSSL = true;
 
   teardown(() => {
@@ -20,7 +21,7 @@ suite('Proxy', () => {
 
   test('No proxy set', () => {
     const getConfiguration = sinon.stub();
-    getConfiguration.withArgs('http', 'proxyStrictSSL').returns(proxyStrictSSL);
+    getConfiguration.withArgs('http', 'proxy').returns(undefined);
 
     const workspace = {
       getConfiguration,
@@ -77,7 +78,7 @@ suite('Proxy', () => {
     assert.deepStrictEqual(agent?.proxy.rejectUnauthorized, proxyStrictSSL);
   });
 
-  test('getProxyEnvVariable should return the proxy as env var', () => {
+  test('getProxyEnvVariable should return the https proxy as env var', () => {
     const getConfiguration = sinon.stub();
     getConfiguration.withArgs('http', 'proxy').returns(proxy);
     getConfiguration.withArgs('http', 'proxyStrictSSL').returns(proxyStrictSSL);
@@ -89,6 +90,6 @@ suite('Proxy', () => {
     const envVariable = getProxyEnvVariable(getProxyOptions(workspace));
 
     // noinspection HttpUrlsUsage
-    assert.deepStrictEqual(envVariable, `http://${auth}@${host}:${port}`);
+    assert.deepStrictEqual(envVariable, `${protocol}//${auth}@${host}:${port}`);
   });
 });